+++ src/canaryusb.h
@@ -23,8 +23,10 @@
  * SOFTWARE.
  */
 
-#define SUBSYSTEM "usb"
-#define DEVICE_TYPE "usb_device"
+#define USB_SUBSYSTEM "usb"
+#define USB_DEVICE_TYPE "usb_device"
+#define SDCARD_SUBSYSTEM "block"
+#define SDCARD_DEVICE_TYPE "disk"
 #define TOTAL_MAX_BASE_32_MESSAGE_LENGTH 118
 #define MAX_BASE_32_MESSAGE_LENGTH 59
 #define CANARY_ON_ACTION "add"
@@ -35,7 +37,7 @@
 #define REPLACE_WITH '_' 
 #define DOT "."
 #define MAX_CANARY_TOKEN_LENGTH 200
-#define MAX_TRUSTED_LIST_LENGTH 1040
+#define MAX_TRUSTED_LIST_LENGTH 2040
 #define CONFIG_FILE ".config/canaryusb/config.toml"
 
 // Argument check types
@@ -51,21 +53,23 @@
 #endif                                   
 
 #define _NAME_ "canaryusb" 
-#define _VERSION_ "3.2.0"
+#define _VERSION_ "4.0.1"
 
 #define BOLD_TEXT "\e[1m"
 #define NO_BOLD_TEXT "\e[m"
 
 #define MAX_PID_LEN 10
 
+extern int dev_fingerprint;
 //These are only for testing reasons.
-extern int usb_fingerprint;
 extern int trusted_list;
 extern char *canary_token;
 extern char *trusted_list_value;
 extern int kill_canaryusb;
+extern int monitor_usb;
+extern int monitor_sdcard;
 
-void monitor_usb();
+void monitor_devices();
 void free_canaries();
 void parse_command_line(int argc, char *argv[]);
 void parse_configuration_file();